Detection methods
Created a new game object (wall) with it's parameters and it's constructor and methods. Also added a counter of walls on the game object container and a method to search for a position in each array of game object references. For this last thing, I also added a get col and get row method in the position class. Using this new method for searching the game objects, I made the whatInPos method in order to complete the positionToString method. I also begun with the isGrounded method and the move one.