Fixed exceptions and fixed game configurations
Exceptions are thrown as wished and catched as wished too.
The game has now two saved levels (one and two) which can be acessed with the reset, which will just make a deep copy of the level we want to go to and put it in the conf of the game. Load just loads the path. Reset works in a way that if we load a path and do a reset with 0 args, we will reset to the last path loaded.